fix: request apply insets when the view is attached to the hierarchy (#117)
## 📜 Description

Call `requestApplyInsetsWhenAttached` when `onApplyWindowInsets`
listener is added.

## 💡 Motivation and Context

It's needed in order to fix a situation, when after reloading JS bundle
first keyboard events are not coming. The problem is reproducible only
in bare project (i. e. no other 3rd party libraries), since if you add
navigation to your app, then `react-native-screens` will "fix" the
problem. But of course it's not reliable to rely on 3rd party libraries
behaviour and the library should work in bare project too.

Fixes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/113#issuecomment-1424217269

## 📢 Changelog

### Android
- `requestApplyInsets` when the view is attached to the hierarchy;

## 🤔 How Has This Been Tested?

Tested manually on:
- Pixel 3 (API 28, emulator);
- Pixel 7 Pro (API 33, real device)

## 📝 Checklist

- [x] CI successfully passed