feat: compress images (#126)
## 📜 Description

Compressed images using `tinypng`. Added an action in CI to detect
graphic files that can be optimized and commit them to attract an
atention.

## 💡 Motivation and Context

Current image size is ~200kb for a single image. On some doc pages
sometimes it's possible to see 3 images on the same page and sometimes
it take some time to load all of them. When the size of images will be
reduced the loading speed will be faster, than before and as a result it
will give better experience for end users.

The builtin image compressor doesn't use as powerful alghorithms as
`tinypng` does, so first of all I compressed all images via `tinypng`
and added `calibre` just an additional check (in case if I forget to
optimize an image this action will leave a comment and I will optimize a
concrete image using `tinypng`).

## 📢 Changelog

### CI
- added `compress-images` job, which will try to optimize images, if
it's possible;

### Docs
- compressed images via `tinypng`;

## 🤔 How Has This Been Tested?

Tested locally on localhost:3000 😁 

## 📸 Screenshots (if appropriate):

You will not notice visual difference.

## 📝 Checklist

- [x] CI successfully passed

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>