refactor: move logic away from manager to `EdgeToEdgeReactViewGroup` (#148)
## 📜 Description

Moved logic from manager to `EdgeToEdgeReactViewGroup`.

## 💡 Motivation and Context

To align with standard approaches and unify codebase.

When I've added `KeyboardGestureView` I incapsulated all logic inside
this view and delegated only method calls to manager. In this PR I've
decided to stick with this approach and rework
`EdgeToEdgeReactViewGroup`.

Just theoretically setting up of callbacks could be moved to
`onAttachedToWindow` method, but I've decided to keep changes as minimal
as possible, so I left it as it is.

## 📢 Changelog

### Android
- move logic away from manager to `EdgeToEdgeReactViewGroup`

## 🤔 How Has This Been Tested?

Tested on Pixel 6 (API 28). Both fabric & paper.

## 📝 Checklist

- [x] CI successfully passed