Proposal: Apply prettier on the whole project files (#313)
## 📜 Description

This is a proposal. When I'm working on the repository, I notice that
the formatter (prettier) update some part of the files even if I don't
changed them. This occurs because some files are not formatted with it,
this PR add a new npm command (`format`) and run it.

I know that this is a lot of changes (especially in the docs markdown
files), feels free to close the PR if you don't want to deal with this.

## 💡 Motivation and Context

This simplifies collaboration on the project, can even been run on
commit using `lint-staged`.

---------

Co-authored-by: kirillzyusko <zyusko.kirik@gmail.com>