fix: missing components mocks (#320)
## 📜 Description

Added mocks for exported components, such as `KeyboardAvoidingView`,
`KeyboardAwareScrollView`, `KeyboardStickyView`.

## 💡 Motivation and Context

The library should support mocks for exported components out of the box.

In an ideal world I'd like to mock dependencies that depends on native
implementation and use a real implementation for components. However
since we mock entire module (`react-native-keyboard-controller`) it
becomes impossible to mock it partially.

Even with `jest.requireActual()` for components import from
`react-native-keyboard-controller` produced `{}` and as a result
`useKeyboardHandler` and other native dependencies were undefined and it
causes errors during component rendering.

Also a controversial topic was how to mock these components - as
`View`/`ScrollView` from RN or as strings. I had at implementation of
other popular repositories and came to conclusion, that it should be
mocked as RN components (as a reference I used
`react-native-gesture-handler`).

Closes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/319

## 📢 Changelog

### Docs
- added new checkbox for PR template (to verify mocks);

### CI
- use latest version of sponsor labeling, because old one uses
deprecated Node 12;

### JS

- added mocks for `KeyboardAvoidingView`, `KeyboardAwareScrollView`,
`KeyboardStickyView`;
- use imports from `react-native-keyboard-controller` in `components`;

## 🤔 How Has This Been Tested?

Added new snapshot tests to Fabric and paper examples.

## 📝 Checklist

- [x] CI successfully passed