feat: `parentScrollViewTarget` detection (#347)
## 📜 Description

Added `parentScrollViewTarget` filed to `useReanimatedFocusedInput`.

## 💡 Motivation and Context

The motivation behind this field is described in
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/336

Here I'd like to focus on implementation details. The implementation is
quite straighforward - we add new field to types/specs/events and we're
ready to go. And this is true, but I'd like to give some implementation
details:
- all search of ScrollView parent I moved to extension (added new
property `parentScrollViewTarget` on iOS/Android to focused input);
- on iOS I need to check whether new arch is enabled (the same as in
https://github.com/kirillzyusko/react-native-keyboard-controller/pull/191)
to get correct tag of the view;

Also this PR optimizes performance in case if you have a lot of
`KeyboardAwareScrollView`'s in stack-navigator - before it would scroll
all these ScrollView whenever keyboard appears, but now it scrolls only
`KeyboardAwareScrollView` that is actually visible 👀

Closes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/336

## 📢 Changelog

### Docs

- added a reference about `parentScrollViewTarget`;

### JS

- added `parentScrollViewTarget` to types declaration;
- added `parentScrollViewTarget` to specs;
- started to use `parentScrollViewTarget` in `KeyboardAwareScrollView`
component;
- added a mock + updated unit tests;

### iOS

- find `parentScrollViewTarget` and send it to JS;

### Android

- find `parentScrollViewTarget` and send it to JS;

## 🤔 How Has This Been Tested?

Tested manually on:
- Pixel 7 Pro (Android 14);
- iPhone 15 Pro (iOS 17.2);

## 📸 Screenshots (if appropriate):


https://github.com/kirillzyusko/react-native-keyboard-controller/assets/22820318/3d0e430a-b284-4038-bddc-1e6462649f4f

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed