feat: `showArrows` property (#379)
## 📜 Description

Added `showArrows` property.

## 💡 Motivation and Context

Sometimes we just want to hide buttons. With current API we have to
override button component (and it's actually very complex process and
it's not reasonable to create such complex solution for such simple
purpose).

So in this PR I added `showArrows` property. It doesn't change current
behavior of the component, but gives an ability to hide these arrows 🙂

## 📢 Changelog

### Docs

- added a note about `showArrows` property;

### JS

- added new `showArrows` property;

## 🤔 How Has This Been Tested?

Tested manually on iPhone 15 Pro (iOS 17.2).

## 📸 Screenshots (if appropriate):


![image](https://github.com/kirillzyusko/react-native-keyboard-controller/assets/22820318/1a2b115c-d445-41e7-85dc-512accc7b9bd)

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed