fix: update Podfile.lock after RN upgrade (#385)
## 📜 Description

Update `Podfile.lock`.

## 💡 Motivation and Context

In
https://github.com/kirillzyusko/react-native-keyboard-controller/pull/299
I updated RN to `0.73.5`. However straight before a merge a new version
has been released, and I updated it to `0.73.6` (but I forgot to remove
Pods directory). As a result on CI we were trying to install `0.73.6`
version (without cache) when `0.73.5` was declared in `Podfile.lock`.

So this is a follow up for
https://github.com/kirillzyusko/react-native-keyboard-controller/pull/299

## 📢 Changelog

### iOS

- update `Podfile.lock`

## 🤔 How Has This Been Tested?

There is only a way to test it on CI 👀 

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed