refactor: `onChangeText` from delegate  (#435)
## 📜 Description

Consume `onChangeText` from delegate level.

## 💡 Motivation and Context

To make some of functionality properly working in reach RN ecosystem you
had to apply patch to `react-native-text-input-mask`.

However it makes it complicated in terms of setup so I always thought on
how to improve it. In
https://github.com/kirillzyusko/react-native-keyboard-controller/pull/426
we started to inject our delegate to intercept some `TextInput` events.
So now we can intercept necessary events without event-bus subscription
(we can take them directly from the delegate).

In this PR I re-worked this part (removed `TextInputObserver` class) and
started to use delegate for these purposes. Also I've discovered again
that we may substitute back delegate and then substitute our own again
(when user typed first letter) - it's not critical but it doesn't comply
with keyboard-controller delegate paradigm - we should set delegate only
once (on focus) and unset it on un-focus. So I also re-worked this part
and now, when we're trying to delete our observers - we check, that we
are actually move focused to another input.

## 📢 Changelog

### Docs

- removed **Known issues** from `useFocusedInputHandler` hook;

### E2E

- compare text between `worklet` and `formatted`;

### iOS

- substitute delegate only one time on focus;
- handle text changes on delegate level;

## 🤔 How Has This Been Tested?

Tested on CI and iPhone 15 Pro manually.

## 📸 Screenshots (if appropriate):

No visual changes.

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed