fix: missing `navigationBar` height in keyboard size if `navigationBar` is `translucent` (#469)
## 📜 Description

Keyboard size is incorrectly reported if we draw behind navigation
status bar (i. e. it's translucent). This PR fixes this by omitting the
`-navigationBarHeight` if `navigationBar` is translucent.

## 💡 Motivation and Context

It's incorrect behavior in the core of the library and should be fixed 👀

Originally discussed in
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/259#issuecomment-1848966075

Maybe there is better solution which don't directly depends on
`navigationBarTrasnlucent` property, but to fix this bug right now let's
use this code. Later on we can always improve/refactor the code.

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### Android

- added `hasTranslucentNavigationBar` to `KeyboardAnimationCallback`;
- conditionally subtract `navigationBar` insets;

## 🤔 How Has This Been Tested?

Tested manually on Pixel 7 Pro and on CI (via e2e tests).

## 📸 Screenshots (if appropriate):

Tested on Pixel 7 Pro (Android 14):

|Before|After|
|-------|-----|

|![telegram-cloud-photo-size-2-5427266154814104062-y](https://github.com/kirillzyusko/react-native-keyboard-controller/assets/22820318/58f14456-5aee-4e32-9181-0238dc82ac85)|![telegram-cloud-photo-size-2-5427266154814104063-y](https://github.com/kirillzyusko/react-native-keyboard-controller/assets/22820318/cbbf0460-af76-42aa-b5b3-ef52db1ebfc7)|

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed