fix: random scrolls of `KeyboardAwareScrollView` during re-render (when keyboard is open) (#477)
## 📜 Description

Fixed random `KeyboardAwareScrollView` scrolls when user types a message
on Android < 11.

## 💡 Motivation and Context

The problem happens when you update a state/props and
`KeyboardAwareScrollView` gets re-rendered. In this case `handler`
object gets re-created and `useAnimatedReaction` produce a new
`onMove`/`onEnd` events.

Since we didn't call `onStart` -> we have out-of-date variables and
because of that scroll happens to undesired position. The fix should be
simple - `onMove` shouldn't be fired. So to fix that I changed
`[handler] -> deps` (actually handler will be rebuild and new version
will be used only when deps were updated -> otherwise new handler will
not be used - this is how `useKeyboardHandler` hook works).

Additionally I cast `deps` to `unknown` to avoid types mismatches
between `REA` and `React`.

Closes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/476
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/384

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### JS

- use `deps` instead of `[handler]` as an array of dependencies;

## 🤔 How Has This Been Tested?

Tested on CI.

### 📸 Screenshots (if appropriate):

|Before|After|
|------|------|

|![image](https://github.com/kirillzyusko/react-native-keyboard-controller/assets/22820318/e5736e1f-5503-4022-9a54-b5b11906a75c)|![image](https://github.com/kirillzyusko/react-native-keyboard-controller/assets/22820318/40342eb9-878b-4012-8d6e-37f567a4c6a3)|

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed